Refactor core into measures and analysis submodules
